backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-storage==2.13.0
google-cloud-documentai==2.20.1
pydantic==2.4.2
python-dotenv==1.0.0
